+++ __tests__/e2e/specs/homepage_spec.js
@@ -352,13 +352,13 @@ describe("Homepage", () => {
           .click();
         cy.get(".HeaderSearch .search-input")
           .first()
-          .type("ARK Bounty{enter}");
+          .type("Genesis Wallet{enter}");
 
         cy.get("h1").should($heading2 => {
           expect($heading2.text()).not.to.eq(heading);
         });
 
-        cy.url().should("include", "/wallets/AYCTHSZionfGoQsRnv5gECEuFWcZXS38gs");
+        cy.url().should("include", "/wallets/AewxfHQobSc49a4radHp74JZCGP8LRe4xA");
       });
     });