+++ src/components/tables/mobile/Transactions.vue
@@ -43,7 +43,7 @@
             {{ $t("TRANSACTION.SMARTBRIDGE") }}
           </div>
           <div class="text-right truncate">
-            {{ emojify(transaction.vendorField) }}
+            {{ sanitizeVendorfield(transaction.vendorField) }}
           </div>
         </div>