+++ src/services/index.ts
@@ -1,31 +1,33 @@
 import ApiService from "./api";
-import BlockService from "./block";
 import BlockchainService from "./blockchain";
+import BlockService from "./block";
 import BridgechainService from "./bridgechain";
 import BusinessService from "./business";
 import CryptoCompareService from "./crypto-compare";
 import DelegateService from "./delegate";
 import ForgingService from "./forging";
+import LockService from "./lock";
 import MigrationService from "./migration";
 import NodeService from "./node";
-import SearchService from "./search";
 import RoundService from "./round";
+import SearchService from "./search";
 import TransactionService from "./transaction";
 import WalletService from "./wallet";
 
 export {
   ApiService,
-  BlockService,
   BlockchainService,
+  BlockService,
   BridgechainService,
   BusinessService,
   CryptoCompareService,
   DelegateService,
   ForgingService,
+  LockService,
   MigrationService,
   NodeService,
-  SearchService,
   RoundService,
+  SearchService,
   TransactionService,
   WalletService,
 };